Increased size of returned frames
The side of each frame was increased to equal the hypotenuse of the
image, to account for the cuttoff of the corners during rotation.

An offsest for centering the image was also calculated and use to form
the final frames.

Additionally, the empty spaces from the ouput have been filled in with
'#' by default instead of ' ' to reflect that jpegs have white
backgrounds be default